Change file name for the function.
